+++ server/src/routes/players.ts
@@ -1,8 +1,17 @@
 import express, { Router } from "express";
 import { importPlayersFromCsv } from "../import/playerCsv";
+import { listPlayers } from "../db/players";
 import { ApiError } from "../http/errors";
 
 export const playersRouter = Router();
+
+playersRouter.get("/", async (_req, res, next) => {
+  try {
+    res.json(await listPlayers());
+  } catch (err) {
+    next(err);
+  }
+});
 
 playersRouter.post(
   "/import",